Separate job API calls from the model.
